Tarea #12
Mantenimiento de UA, edificios y personal.

+++ rolsac2-persistence/src/main/java/es/caib/rolsac2/persistence/model/pk/JUsuarioUnidadAdministrativaPK.java
@@ -1,42 +1,45 @@
 package es.caib.rolsac2.persistence.model.pk;
 
-import javax.persistence.Column;
-import javax.persistence.Embeddable;
 import java.io.Serializable;
 import java.util.Objects;
+
+import javax.persistence.Column;
+import javax.persistence.Embeddable;
 
 @Embeddable
 public class JUsuarioUnidadAdministrativaPK implements Serializable {
     private static final long serialVersionUID = -4971184161239060579L;
     @Column(name = "UAUS_CODUSER", nullable = false)
-    private Integer usuario;
+    private Long usuario;
 
     @Column(name = "UAUS_CODUA", nullable = false)
-    private Integer unidadAdministrativa;
+    private Long unidadAdministrativa;
 
-    public Integer getUsuario() {
+    public Long getUsuario() {
         return usuario;
     }
 
-    public void setUsuario(Integer uausCoduser) {
+    public void setUsuario(Long uausCoduser) {
         this.usuario = uausCoduser;
     }
 
-    public Integer getUnidadAdministrativa() {
+    public Long getUnidadAdministrativa() {
         return unidadAdministrativa;
     }
 
-    public void setUnidadAdministrativa(Integer uausCodua) {
+    public void setUnidadAdministrativa(Long uausCodua) {
         this.unidadAdministrativa = uausCodua;
     }
 
     @Override
     public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || this.getClass() != o.getClass()) return false;
+        if (this == o)
+            return true;
+        if (o == null || this.getClass() != o.getClass())
+            return false;
         JUsuarioUnidadAdministrativaPK entity = (JUsuarioUnidadAdministrativaPK) o;
-        return Objects.equals(this.usuario, entity.usuario) &&
-                Objects.equals(this.unidadAdministrativa, entity.unidadAdministrativa);
+        return Objects.equals(this.usuario, entity.usuario)
+                        && Objects.equals(this.unidadAdministrativa, entity.unidadAdministrativa);
     }
 
     @Override
